Controle de regras de negocio
Criacao de todas as camadas para Perfil e PessoaPerfil

Criacao de parsers para uso futuro

Correcao de imports e alteracoes de formatacao em quase todas as tabelas

Codigo pendente de testes unitarios

+++ src/main/java/com/stefanini/servico/PerfilServico.java
@@ -1,14 +1,23 @@
 package com.stefanini.servico;
 
+import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import javax.ejb.Stateless;
+import javax.ejb.TransactionAttribute;
+import javax.ejb.TransactionAttributeType;
+import javax.ejb.TransactionManagement;
+import javax.ejb.TransactionManagementType;
 import javax.inject.Inject;
+import javax.transaction.Transactional;
 import javax.validation.Valid;
 
 import com.stefanini.dao.PerfilDAO;
+import com.stefanini.dto.PerfilDTO;
 import com.stefanini.model.Perfil;
-import com.stefanini.util.IGenericService;
+import com.stefanini.parser.PerfilParser;
 
 /**
  * 
@@ -16,38 +25,63 @@ import com.stefanini.util.IGenericService;
  * @author yuri araujo de castro silva
  *
  */
-public class PerfilServico implements IGenericService<Perfil, Long> {
+@Stateless
+@TransactionManagement(TransactionManagementType.CONTAINER)
+@TransactionAttribute(TransactionAttributeType.NOT_SUPPORTED)
+@Transactional
+public class PerfilServico implements Serializable {
 	
-	@Inject
+	/**
+   * serial da classe
+   */
+  private static final long serialVersionUID = 1L;
+  
+  @Inject
 	private PerfilDAO dao;
 	
+  @Inject
+  private PerfilParser perfilParser; 
+  
 	/**
 	 * Salvar os dados de um Perfil
 	 */
+  @TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public Perfil salvar(@Valid Perfil perfil) {
+    if(dao.isNomeRepeated(perfil.getNome())) {
+      return new Perfil();
+    }
 		return dao.salvar(perfil);
 	}
 
 	/**
 	 * Atualizar o dados de um Perfil
 	 */
-	@Override
-	public Perfil atualizar(@Valid Perfil entity) {
-		return dao.atualizar(entity);
+  @TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
+	public Perfil atualizar(@Valid Perfil perfil) {
+    if(dao.isNomeRepeated(perfil.getNome())) {
+      return new Perfil();
+    }
+		return dao.atualizar(perfil);
 	}
 
 	/**
 	 * Remover um Perfil pelo id
 	 */
-	@Override
-	public void remover(Long id) {
-		dao.remover(id);		
+  @TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
+	public Boolean remover(Long id) {
+    Optional<Perfil> perfil = dao.encontrar(id);
+    if(perfil.isPresent()) {
+      if(perfil.get().getPessoas().isEmpty()) {
+        dao.remover(id);
+        return true;
+      }
+    }
+		return false;
 	}
 
 	/**
 	 * Buscar uma lista de Perfil
 	 */
-	@Override
 	public Optional<List<Perfil>> getList() {
 		return dao.getList();
 	}
@@ -55,9 +89,17 @@ public class PerfilServico implements IGenericService<Perfil, Long> {
 	/**
 	 * Buscar um Perfil pelo ID
 	 */
-	@Override
 	public Optional<Perfil> encontrar(Long id) {
 		return dao.encontrar(id);
 	}
+
+  public Optional<List<PerfilDTO>> encontrarPorNome(String nome) {
+    Optional<List<Perfil>> optPerfil = dao.encontrarPorNome(nome);
+    if(optPerfil.isPresent()) {
+      return Optional.of(perfilParser.toDTOList(optPerfil.get()));
+    }
+    List<PerfilDTO> retorno = new ArrayList<PerfilDTO>();
+    return Optional.of(retorno);
+  }
 
 }
